commit de recorridototal

+++ src/main/java/org/openjfx/Viaje.java
@@ -6,7 +6,7 @@ public class Viaje {
     private String transporte;
     private double capacidadGasolina;
     private double consumoGasolina;
-    private int recorridoTotal = 0;
+    private double recorridoTotal = 0;
 
     public Viaje(String puntoInicial, String puntoFinal, String transporte) {
         this.puntoInicial = puntoInicial;
@@ -54,11 +54,11 @@ public class Viaje {
         public void setConsumoGasolina(double consumoGasolina) {
             this.consumoGasolina = consumoGasolina;
         }
-        public int getRecorridoTotal() {
+        public double getRecorridoTotal() {
             return recorridoTotal;
         }
 
-        public void incrementarRecorridoTotal(int incremento) {
+        public void incrementarRecorridoTotal(double incremento) {
             this.recorridoTotal += incremento;
         }
     @Override
